move benchmarks  to nested project
