+++ bench/revm_comparison/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "revm_factorial"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+evm_mlir = { path = "../../" }
+hex = "0.4.3"
+revm = "9.0.0"
+
+[[bin]]
+name="evm_mlir_factorial"
+path="src/evm_mlir_factorial.rs"
+
+[[bin]]
+name="revm_factorial"
+path="src/revm_factorial.rs"
